Resolution scaling saving, snapshotting improvements
Changes:
- Added Metadata plist to document folder
- Stored selected resolution scale factor in metadata
- Added save / load for res scale so documents resume at the last used resolution
- Added time-based frame delta instead of fixed 1/60
- Changed snapshotting to use drawViewHierarchyInRect for simplicity (also fixes a crash when snapshotting)

I'd say STWSnappshotter can be removed, as it's not really needed. Only downside of the new setup is that shapshots aren't at the correct aspect ratio, but that could be fixed by snapshotting at the same aspect ratio as the actual view but at a smaller size, then showing the thumbnail as aspectFill.